Ignore broken workspace test
It looks like @mcgoo was adding the ability to specify vcpkg
dependencies at the workspace manifest level.
combine_deps_from_all_crates_and_workspace is a new test that is meant
to cover this case, though as is it doesn't work.

Looking more closely, I see that the values nested under 'xworkspace'
are getting dropped by `cargo metadata` completely.  Changing these to
be in the 'workspace' namespace makes them show up in the metadata
output; however, to process that metadata, presumably we still need to
have a root crate that indicates which git rev to use for the vcpkg tree.

Given that currently pointing cargo vcpkg at a workspace base is
currently incomplete (see mcgoo/cargo-vcpkg#1 and
run_on_workspace_fails()), disabling the test for now seems preferable
in order to allow a release.